    % File /man/test_data.Rd
     \name{test_data}
     \docType{data}
     \alias{test_data}
     \title{Sample Data for Testing Purposes}
     \description{
       This table contains the first 100 records of iRefIndex/MITAB for E.coli. It is just a small table to be used for testing purposes.
     }
     \usage{data(test_data)}
     \format{A table with 100 rows and 54 columns.}
     \source{Generated from iRefIndex using iRefR.}
     \keyword{datasets}
